Include some details of the tooling API request in the configuration cache key generated to store the result of the request.
A dummy key is generated for build action requests, as the action is currently not exposed by the tooling API client.
